+++ CHANGES.rst
@@ -6,6 +6,8 @@ Unreleased
 -   Drop support for Python 3.9. :pr:`5730`
 -   Remove previously deprecated code: ``__version__``. :pr:`5648`
 
+-   Fix template_filter decorator to correctly register filters and support usage with or without parentheses.
+-   Improve docstring for `template_filter` decorator for better clarity and examples.
 
 Version 3.1.1
 -------------